Fix: Bookmark.ts -> BookmarkEntryList.ts、BookmarkCompatibilityLayer.ts -> Bookmark.tsに改名
getAllThreads(), getAllBoards()をapp.Bookmarkにも実装
